Feature/add synchronization

+++ lib/boxxspring/worker/base.rb
@@ -60,10 +60,10 @@ module Boxxspring
             payload = self.payload_from_message( message )
             if payload.present?
               begin
-                self.process_payload( payload )
+                result = self.process_payload( payload )
                 # note: if an exception is raised the message will not be 
                 #       deleted
-                self.delete_message( message )
+                self.delete_message( message ) unless result == false
               rescue StandardError => error
                 self.logger.error(
                   "The #{ self.human_name } worker failed to process the " +
